Since it is not possible to use computed gotos in Visual Studio/VC++ I have simulated the dispatch table with a switch case containing goto statements.