@munificent 2015-06-27
Interesting!

How does this affect the performance? Can you run the benchmarks with and without this change and show us what the numbers look like?