@BarabasGitHub 2015-06-29
There are some numbers in the discussion on issue #269
If you want I can try to run the benchmarks with the old and new approach later this week.