@BarabasGitHub 2015-07-08
Sorry I'm a bit late, but some things came up. In any case I ran the benchmark script, with the original to generate a base-line and then with the new implementation.

Original:
```
binary_trees - wren            .......... 0.49s 0.0092 no baseline
delta_blue - wren              .......... 0.23s 0.0036 no baseline
fib - wren                     .......... 0.62s 0.0027 no baseline
for - wren                     .......... 0.31s 0.0017 no baseline
method_call - wren             .......... 0.41s 0.0100 no baseline
map_numeric - wren             .......... 0.80s 0.0133 no baseline
map_string - wren              .......... 0.20s 0.0024 no baseline
string_equals - wren           .......... 0.68s 0.0045 no baseline
```
New:
```
binary_trees - wren            .......... 0.42s 0.0048 118.27% relative to baseline
delta_blue - wren              .......... 0.19s 0.0013 121.51% relative to baseline
fib - wren                     .......... 0.55s 0.0033 113.09% relative to baseline
for - wren                     .......... 0.26s 0.0030 122.18% relative to baseline
method_call - wren             .......... 0.29s 0.0018 138.31% relative to baseline
map_numeric - wren             .......... 0.69s 0.0042 115.32% relative to baseline
map_string - wren              .......... 0.18s 0.0016 107.73% relative to baseline
string_equals - wren           .......... 0.54s 0.0027 125.09% relative to baseline
```