@munificent 2015-07-18
OK, I'm sold. :)

I wish we didn't have to turn wren_opcodes.h into a giant macro, but I can't think of any other solution.

I'll add some other comments at specific points, but would you mind also rebasing this patch against the latest master? I think some stuff has moved under you, making this hard to merge in.

Thank you for doing so much work on this!