@BarabasGitHub 2015-07-18
I also really dislike the wren_opcodes macro. I wish it could be different.

Anyway, I tried to rebase the goto branch onto the new master branch. But somewhere along the line I did something that made my master branch end up being the same as the goto branch. I think I might have pushed it to the wrong branch or so... AND I fixed it! 

I didn't change your suggestions. Should I do that?