@paulsapps 2016-05-11
Did you try what a function pointer table to declspec(naked) functions looks like with MSVC? Since you can get the address of the function thats like taking the address of a label, then the "naked" function call should become a jmp?