@BarabasGitHub 2016-05-11
I don't think that will work. It says on msdn: "The *naked* attribute
affects only the nature of the compiler's code generation for the
function's prolog and epilog sequences. It does not affect the code that is
generated for calling such functions. Thus, the *naked* attribute is not
considered part of the function's type, and function pointers cannot have
the *naked* attribute." (
https://msdn.microsoft.com/en-us/library/21d5kd3a.aspx)

On 11 May 2016 at 09:18, Paul <notifications@github.com> wrote:

> Did you try what a function pointer table to declspec(naked) functions
> looks like with MSVC? Since you can get the address of the function thats
> like taking the address of a label, then the "naked" function call should
> become a jmp?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/munificent/wren/pull/272#issuecomment-218381130>
>