@paulsapps 2016-06-17
I saw something else related to this today "__assume" https://msdn.microsoft.com/en-us/library/1b3fsfxw.aspx apparently you could use this on the switch value to stop it generating the extract checks. Might be worth checking if you ever get time.