@munificent 2017-03-25
Ugh, I'm so sorry, but I left this open so long that I'm not able to merge it in now since so many changes have happened since. :(